Fix granule discovery for files without file type config
A granule is now included in discovery results even when none of its
files has a matching file type in the associated collection
configuration.  Previously, if all files for a granule were unmatched by
a file type configuration, the granule was excluded from the discovery
results.  Now, instead, such a granule is included in the discovery
results, but with an empty file list.

Added corresponding unit and integration tests for this fix.

Also addressed 3 ESLint warnings for JSDoc violations on the
`addUrlPathToGranuleFiles` and `addUniqueGranuleFilePathToGranuleFiles`
functions in `example/spec/helpers/granuleUtils.js`, along with some
refactoring of those functions to eliminate unnecessary (excess) object
cloning.

CUMULUS-1572