Define property to ignore collection files config
Added support for the `boolean` property `ignoreFilesConfig` at both the
task config level for the `DiscoverGranules` task, as well as at the
collection configuration level for overriding the task-level value.
By default, the property is `false`, meaning that during granule
discovery, a file for a granule is ingested only when the name of the
file matches one of the regexes in the collection's `files` config,
which is the current behavior.

However, setting this property to `true` on the `DiscoverGranules` task
will cause all `files` configs for all collections to be ignored, such
that all files for a granule will be ingested (because there will be no
filename checks against the regexes in the `files` config).

However, it is possible to override the property value set on the task
by setting the property on individual collections.

CUMULUS-1572